import fixes + fix import loop + fix minor signing things
