+++ test_libs/pyspec/eth2spec/test/helpers/genesis.py
@@ -9,7 +9,7 @@ from eth2spec.utils.minimal_ssz import hash_tree_root
 def build_mock_validator(i: int, balance: int):
     pubkey = pubkeys[i]
     # insecurely use pubkey as withdrawal key as well
-    withdrawal_credentials = spec.BLS_WITHDRAWAL_PREFIX_BYTE + hash(pubkey)[1:]
+    withdrawal_credentials = spec.BLS_WITHDRAWAL_PREFIX_BYTE + spec.hash(pubkey)[1:]
     return spec.Validator(
         pubkey=pubkeys[i],
         withdrawal_credentials=withdrawal_credentials,
