+++ test_libs/pyspec/eth2spec/test/helpers/state.py
@@ -1,8 +1,7 @@
 # Access constants from spec pkg reference.
 import eth2spec.phase0.spec as spec
 
-from eth2spec.phase0.state_transition import state_transition_to, state_transition
-from eth2spec.test.helpers.block import build_empty_block
+from eth2spec.phase0.state_transition import state_transition_to
 
 
 def get_balance(state, index):
@@ -22,16 +21,6 @@ def next_epoch(state):
     """
     slot = state.slot + spec.SLOTS_PER_EPOCH - (state.slot % spec.SLOTS_PER_EPOCH)
     state_transition_to(state, slot)
-
-
-def apply_empty_block(state):
-    """
-    Transition via an empty block (on current slot, assuming no block has been applied yet).
-    :return: the empty block that triggered the transition.
-    """
-    block = build_empty_block(state)
-    state_transition(state, block)
-    return block
 
 
 def get_state_root(state, slot) -> bytes:
